Use insta for snapshot testing
