--- samples/event_with_entity_ref.xml
@@ -1,81 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Event xmlns="http://schemas.microsoft.com/win/2004/08/events/event">
-    <System>
-        <Provider Name="Microsoft-Windows-CAPI2" Guid="{5bbca4a8-b209-48dc-a8c7-b23d3e5216fb}">
-        </Provider>
-        <EventID>53</EventID>
-        <Version>0</Version>
-        <Level>4</Level>
-        <Task>53</Task>
-        <Opcode>2</Opcode>
-        <Keywords>0x4000000000000036</Keywords>
-        <TimeCreated SystemTime="2017-05-19 02:02:36.203125 UTC">
-        </TimeCreated>
-        <EventRecordID>28</EventRecordID>
-        <Correlation>
-        </Correlation>
-        <Execution ProcessID="1396" ThreadID="2132">
-        </Execution>
-        <Channel>Microsoft-Windows-CAPI2/Operational</Channel>
-        <Computer>WIN-M5327EF98B9</Computer>
-        <Security UserID="S-1-5-21-1223297778-3299746493-1462173606-500">
-        </Security>
-    </System>
-    <UserData>
-        <CryptRetrieveObjectByUrlWire>
-            <URL scheme="http">http://crl.microsoft.com/pki/crl/products/microsoftrootcert.crl</URL>
-            <Object type="CONTEXT_OID_CRL" constant="2">
-            </Object>
-            <Timeout>PT9.984S</Timeout>
-            <Flags value="202005" CRYPT_RETRIEVE_MULTIPLE_OBJECTS="true" CRYPT_WIRE_ONLY_RETRIEVAL="true" CRYPT_LDAP_SCOPE_BASE_ONLY_RETRIEVAL="true" CRYPT_PROXY_CACHE_RETRIEVAL="true">
-            </Flags>
-            <AuxInfo maxUrlRetrievalByteCount="104857600" fProxyCacheRetrieval="true">
-            </AuxInfo>
-            <AdditionalInfo>
-                <NetworkConnectivityStatus value="1" _SENSAPI_NETWORK_ALIVE_LAN="true">
-                </NetworkConnectivityStatus>
-                <Action name="NoProxy">
-                </Action>
-                <Action name="NoProxy">
-                </Action>
-                <HTTPRequestHeadersInfo>
-                    <Header>GET /pki/crl/products/microsoftrootcert.crl HTTP/1.1</Header>
-                    <Header>Accept: */*</Header>
-                    <Header>If-None-Match: &amp;quotea9ee7b1bc43d21:0&amp;quot</Header>
-                    <Header>If-Modified-Since: Mon, 21 Nov 2016 06:01:26 GMT</Header>
-                    <Header>Cache-Control: max-age = 900</Header>
-                    <Header>User-Agent: Microsoft-CryptoAPI/6.1</Header>
-                    <Header>Connection: Keep-Alive</Header>
-                </HTTPRequestHeadersInfo>
-                <HTTPResponseHeadersInfo>
-                    <Header>HTTP/1.1 200 OK</Header>
-                    <Header>Connection: keep-alive</Header>
-                    <Header>Date: Thu, 18 May 2017 11:37:58 GMT</Header>
-                    <Header>Content-Length: 813</Header>
-                    <Header>Content-Type: application/pkix-crl</Header>
-                    <Header>Last-Modified: Tue, 02 May 2017 22:24:24 GMT</Header>
-                    <Header>ETag: 0x8D491A9FD112A27</Header>
-                    <Header>Server: Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0</Header>
-                    <Header>x-ms-request-id: 477c132d-0001-0045-443b-c49ae1000000</Header>
-                    <Header>x-ms-version: 2009-09-19</Header>
-                    <Header>x-ms-lease-status: unlocked</Header>
-                    <Header>x-ms-blob-type: BlockBlob</Header>
-                </HTTPResponseHeadersInfo>
-            </AdditionalInfo>
-            <CacheInfo lastSyncTime="2017-05-19T02:02:36.203Z">
-                <URLCacheResponseInfo responseType="CRYPTNET_URL_CACHE_RESPONSE_HTTP" lastModifiedTime="2017-05-02T22:24:24Z">
-                </URLCacheResponseInfo>
-            </CacheInfo>
-            <RetrievedObjects>
-                <CertificateRevocationList fileRef="0986764ED95D3C77F3F1AD8340EDD2F36C3BF8E7.crl" issuerName="Microsoft Root Certificate Authority">
-                </CertificateRevocationList>
-            </RetrievedObjects>
-            <EventAuxInfo ProcessName="Setup.exe">
-            </EventAuxInfo>
-            <CorrelationAuxInfo TaskId="{74E4CD40-C966-49F5-B50A-032DFFEE57CA}" SeqNumber="7">
-            </CorrelationAuxInfo>
-            <Result value="0">
-            </Result>
-        </CryptRetrieveObjectByUrlWire>
-    </UserData>
-</Event>
