--- samples/event_with_entity_ref_2.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Event xmlns="http://schemas.microsoft.com/win/2004/08/events/event">
-    <System>
-        <Provider Name="Microsoft-Windows-CAPI2" Guid="{5bbca4a8-b209-48dc-a8c7-b23d3e5216fb}">
-        </Provider>
-        <EventID>51</EventID>
-        <Version>0</Version>
-        <Level>4</Level>
-        <Task>51</Task>
-        <Opcode>2</Opcode>
-        <Keywords>0x400000000000002e</Keywords>
-        <TimeCreated SystemTime="2017-05-19 02:02:36.187500 UTC">
-        </TimeCreated>
-        <EventRecordID>25</EventRecordID>
-        <Correlation>
-        </Correlation>
-        <Execution ProcessID="1396" ThreadID="2132">
-        </Execution>
-        <Channel>Microsoft-Windows-CAPI2/Operational</Channel>
-        <Computer>WIN-M5327EF98B9</Computer>
-        <Security UserID="S-1-5-21-1223297778-3299746493-1462173606-500">
-        </Security>
-    </System>
-    <UserData>
-        <CryptRetrieveObjectByUrlCache>
-            <URL scheme="http">http://crl.microsoft.com/pki/crl/products/microsoftrootcert.crl</URL>
-            <Object type="CONTEXT_OID_CRL" constant="2">
-            </Object>
-            <Flags value="2003" CRYPT_RETRIEVE_MULTIPLE_OBJECTS="true" CRYPT_CACHE_ONLY_RETRIEVAL="true" CRYPT_LDAP_SCOPE_BASE_ONLY_RETRIEVAL="true">
-            </Flags>
-            <AuxInfo maxUrlRetrievalByteCount="104857600">
-            </AuxInfo>
-            <CacheInfo lastSyncTime="2017-02-03T20:08:47.908Z">
-                <URLCachePrefetchInfo objectType="CRYPTNET_URL_CACHE_PRE_FETCH_CRL" thisUpdateTime="2016-11-20T21:48:50Z" nextUpdateTime="2017-02-19T10:08:50Z" publishTime="2017-02-18T21:58:50Z">
-                </URLCachePrefetchInfo>
-                <URLCacheFlushInfo expireTime="2017-02-19T10:08:50Z">
-                </URLCacheFlushInfo>&amp;quot&amp;quot<URLCacheResponseInfo responseType="CRYPTNET_URL_CACHE_RESPONSE_HTTP" responseValidated="true" lastModifiedTime="2016-11-21T06:01:26Z" maxAge="900" eTag="ea9ee7b1bc43d21:0">
-            </URLCacheResponseInfo>
-            </CacheInfo>
-            <RetrievedObjects>
-                <CertificateRevocationList fileRef="B2A57B544F9B889695FC28CFE5EEDAA85B4B6602.crl" issuerName="Microsoft Root Certificate Authority">
-                </CertificateRevocationList>
-            </RetrievedObjects>
-            <EventAuxInfo ProcessName="Setup.exe">
-            </EventAuxInfo>
-            <CorrelationAuxInfo TaskId="{74E4CD40-C966-49F5-B50A-032DFFEE57CA}" SeqNumber="4">
-            </CorrelationAuxInfo>
-            <Result value="0">
-            </Result>
-        </CryptRetrieveObjectByUrlCache>
-    </UserData>
-</Event>
