--- samples/event_with_eventdata.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Event xmlns="http://schemas.microsoft.com/win/2004/08/events/event">
-    <System>
-        <Provider Name="Microsoft-Windows-Security-Auditing" Guid="54849625-5478-4994-A5BA-3E3B0328C30D">
-        </Provider>
-        <EventID>4688</EventID>
-        <Version>2</Version>
-        <Level>0</Level>
-        <Task>13312</Task>
-        <Opcode>0</Opcode>
-        <Keywords>0x8020000000000000</Keywords>
-        <TimeCreated SystemTime="2018-07-28 07:24:45.754787 UTC">
-        </TimeCreated>
-        <EventRecordID>1</EventRecordID>
-        <Correlation>
-        </Correlation>
-        <Execution ProcessID="4" ThreadID="32">
-        </Execution>
-        <Channel>Security</Channel>
-        <Computer>WIN-LL0C19JS506</Computer>
-        <Security>
-        </Security>
-    </System>
-    <EventData>
-        <Data Name="SubjectUserSid">S-1-5-18</Data>
-        <Data Name="SubjectUserName">-</Data>
-        <Data Name="SubjectDomainName">-</Data>
-        <Data Name="SubjectLogonId">0x3e7</Data>
-        <Data Name="NewProcessId">0x58</Data>
-        <Data Name="NewProcessName">Registry</Data>
-        <Data Name="TokenElevationType">%%1936</Data>
-        <Data Name="ProcessId">0x4</Data>
-        <Data Name="CommandLine"></Data>
-        <Data Name="TargetUserSid">S-1-0-0</Data>
-        <Data Name="TargetUserName">-</Data>
-        <Data Name="TargetDomainName">-</Data>
-        <Data Name="TargetLogonId">0x0</Data>
-        <Data Name="ParentProcessName"></Data>
-        <Data Name="MandatoryLabel">S-1-16-16384</Data>
-    </EventData>
-</Event>
