--- samples/event_with_multiple_nodes_same_name_separate_attr.json
@@ -1,114 +1,0 @@
-{
-    "Event": {
-      "System": {
-        "Channel": "Microsoft-Windows-CAPI2/Operational",
-        "Computer": "WIN-M5327EF98B9",
-        "Correlation": null,
-        "EventID": 53,
-        "EventRecordID": 28,
-        "Execution_attributes": {
-          "ProcessID": 1396,
-          "ThreadID": 2132
-        },
-        "Keywords": "0x4000000000000036",
-        "Level": 4,
-        "Opcode": 2,
-        "Provider_attributes": {
-          "Guid": "{5bbca4a8-b209-48dc-a8c7-b23d3e5216fb}",
-          "Name": "Microsoft-Windows-CAPI2"
-        },
-        "Security_attributes": {
-          "UserID": "S-1-5-21-1223297778-3299746493-1462173606-500"
-        },
-        "Task": 53,
-        "TimeCreated_attributes": {
-          "SystemTime": "2017-05-19T02:02:36.203125Z"
-        },
-        "Version": 0
-      },
-      "UserData": {
-        "CryptRetrieveObjectByUrlWire": {
-          "AdditionalInfo": {
-            "Action_attributes": {
-              "name": "NoProxy"
-            },
-            "HTTPRequestHeadersInfo": {
-              "Header": "Connection: Keep-Alive",
-              "Header_1": "GET /pki/crl/products/microsoftrootcert.crl HTTP/1.1",
-              "Header_2": "Accept: */*",
-              "Header_3": "\"",
-              "Header_4": "If-Modified-Since: Mon, 21 Nov 2016 06:01:26 GMT",
-              "Header_5": "Cache-Control: max-age = 900",
-              "Header_6": "User-Agent: Microsoft-CryptoAPI/6.1"
-            },
-            "HTTPResponseHeadersInfo": {
-              "Header": "x-ms-blob-type: BlockBlob",
-              "Header_1": "HTTP/1.1 200 OK",
-              "Header_10": "x-ms-version: 2009-09-19",
-              "Header_11": "x-ms-lease-status: unlocked",
-              "Header_2": "Connection: keep-alive",
-              "Header_3": "Date: Thu, 18 May 2017 11:37:58 GMT",
-              "Header_4": "Content-Length: 813",
-              "Header_5": "Content-Type: application/pkix-crl",
-              "Header_6": "Last-Modified: Tue, 02 May 2017 22:24:24 GMT",
-              "Header_7": "ETag: 0x8D491A9FD112A27",
-              "Header_8": "Server: Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0",
-              "Header_9": "x-ms-request-id: 477c132d-0001-0045-443b-c49ae1000000"
-            },
-            "NetworkConnectivityStatus_attributes": {
-              "_SENSAPI_NETWORK_ALIVE_LAN": "true",
-              "value": "1"
-            }
-          },
-          "AuxInfo_attributes": {
-            "fProxyCacheRetrieval": "true",
-            "maxUrlRetrievalByteCount": "104857600"
-          },
-          "CacheInfo": {
-            "URLCacheResponseInfo_attributes": {
-              "lastModifiedTime": "2017-05-02T22:24:24Z",
-              "responseType": "CRYPTNET_URL_CACHE_RESPONSE_HTTP"
-            }
-          },
-          "CacheInfo_attributes": {
-            "lastSyncTime": "2017-05-19T02:02:36.203Z"
-          },
-          "CorrelationAuxInfo_attributes": {
-            "SeqNumber": "7",
-            "TaskId": "{74E4CD40-C966-49F5-B50A-032DFFEE57CA}"
-          },
-          "EventAuxInfo_attributes": {
-            "ProcessName": "Setup.exe"
-          },
-          "Flags_attributes": {
-            "CRYPT_LDAP_SCOPE_BASE_ONLY_RETRIEVAL": "true",
-            "CRYPT_PROXY_CACHE_RETRIEVAL": "true",
-            "CRYPT_RETRIEVE_MULTIPLE_OBJECTS": "true",
-            "CRYPT_WIRE_ONLY_RETRIEVAL": "true",
-            "value": "202005"
-          },
-          "Object_attributes": {
-            "constant": "2",
-            "type": "CONTEXT_OID_CRL"
-          },
-          "Result_attributes": {
-            "value": "0"
-          },
-          "RetrievedObjects": {
-            "CertificateRevocationList_attributes": {
-              "fileRef": "0986764ED95D3C77F3F1AD8340EDD2F36C3BF8E7.crl",
-              "issuerName": "Microsoft Root Certificate Authority"
-            }
-          },
-          "Timeout": "PT9.984S",
-          "URL": "http://crl.microsoft.com/pki/crl/products/microsoftrootcert.crl",
-          "URL_attributes": {
-            "scheme": "http"
-          }
-        }
-      }
-    },
-    "Event_attributes": {
-      "xmlns": "http://schemas.microsoft.com/win/2004/08/events/event"
-    }
-  }
