--- samples/event_with_text_and_attributes.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Event xmlns="http://schemas.microsoft.com/win/2004/08/events/event">
-    <System>
-        <Provider Name="EventLog">
-        </Provider>
-        <EventID Qualifiers="32768">6009</EventID>
-        <Level>4</Level>
-        <Task>0</Task>
-        <Keywords>0x80000000000000</Keywords>
-        <TimeCreated SystemTime="2017-07-12 17:16:28.214161 UTC">
-        </TimeCreated>
-        <EventRecordID>1</EventRecordID>
-        <Channel>System</Channel>
-        <Computer>WIN-P4SIAA0SQCO</Computer>
-        <Security>
-        </Security>
-    </System>
-    <EventData>
-        <Data>10.00.,15063,,Multiprocessor Free,0</Data>
-        <Binary></Binary>
-    </EventData>
-</Event>
