--- samples/security_event_1.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Event xmlns="http://schemas.microsoft.com/win/2004/08/events/event">
-    <System>
-        <Provider Name="Microsoft-Windows-Security-Auditing" Guid="54849625-5478-4994-A5BA-3E3B0328C30D">
-        </Provider>
-        <EventID>4608</EventID>
-        <Version>0</Version>
-        <Level>0</Level>
-        <Task>12288</Task>
-        <Opcode>0</Opcode>
-        <Keywords>0x8020000000000000</Keywords>
-        <TimeCreated SystemTime="2016-07-08 18:12:51.681640 UTC">
-        </TimeCreated>
-        <EventRecordID>1</EventRecordID>
-        <Correlation>
-        </Correlation>
-        <Execution ProcessID="456" ThreadID="460">
-        </Execution>
-        <Channel>Security</Channel>
-        <Computer>37L4247F27-25</Computer>
-        <Security>
-        </Security>
-    </System>
-    <EventData>
-    </EventData>
-</Event>
